% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bestMpTuneModel_methods.R
\name{bestMpTuneModel}
\alias{bestMpTuneModel}
\alias{checkConsistency}
\alias{predict.bestMpTuneModel}
\alias{predict.bestMpTuneModel.formula}
\alias{print.bestMpTuneModel}
\alias{print.resampledMpTune}
\title{methods for 'bestMpTuneModel', 'resampleMpTune'}
\usage{
\method{predict}{bestMpTuneModel.formula}(object, newdata, type = c("raw",
  "prob"), ...)

\method{predict}{bestMpTuneModel}(object, newdata, type = c("raw", "prob"),
  ...)

print.bestMpTuneModel(x, digits = 4, row.names = FALSE, ...)

print.resampledMpTune(x, digits = 4, ...)

checkConsistency(object)
}
\arguments{
\item{object}{A 'bestMpTuneModel' or 'resampleMpTune'}

\item{newdata}{A new design matrix if object is 'bestMpTuneModel' or a new data frame if object is 'bestMpTuneModel.formula'}

\item{type}{A string indicated type of prediction, only for classification and possibly survival.}

\item{...}{Just for campatibility}

\item{x}{A 'bestMpTuneModel' or 'resampleMpTune'}

\item{digits}{Number of effective digits to keep for numeric variables}

\item{row.names}{Should row names / numbers be printed}
}
\value{
prediction vector or matrix, performance summary, consistency indicators
}
\description{
predict, print and resample consistency checking methods for 'bestMpTuneModel', 'resampledMpTune'
}
\details{
For predict, it depends on problem types and models,
\itemize{
	\item{'classifiction'} {
		a vector of predictions if type = 'raw' and a matrix where each coloumn is predicted
		probabilities for an each class named by its class label if type = 'prob}
	\item{'regression'} {it is always a vector of prediction}
	\item{'survival'} {
		if type = 'raw', return a vector of 'risk' or predicted survival time (depending on models).
		If type = 'prob', for some model (e.g., randomForestSRC::rfsrc), it returns a matrix of survival probabilities
		with each columns named by survival times}
	}.
Function \code{checkConsistency} checks on the consistency of resampling in 'resampledMpTune
}
\seealso{
See \code{\link{mpTune}} for examples
}

