% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/comine_mpTune.R
\name{combine.mpTune}
\alias{combine.mpTune}
\title{combine mpTune objects}
\usage{
combine.mpTune(...)
}
\arguments{
\item{...}{mpTune objects}
}
\value{
A combined mpTune objects
}
\description{
combine mpTune objects
}

