% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/createModel.R
\name{createModel}
\alias{createModel}
\title{fit a model}
\usage{
createModel(x, y, wts, method, tuneValue, obsLevels, pp = NULL,
  last = FALSE, classProbs, ...)
}
\arguments{
\item{x}{Design matrix}

\item{y}{Response vector}

\item{wts}{Weights of samples}

\item{method}{Model information. See \code{\link{getModelInfo}}}

\item{tuneValue}{A one row data frame containing tuning parameter values}

\item{obsLevels}{Class labels for classification}

\item{pp}{PreProcess object}

\item{last}{A logical for whether the current fit is the final fit}

\item{classProbs}{Should classProbs be computed}

\item{...}{Additional paramters other than tuning parameters passed to model fitting methods}
}
\value{
A list of model fitting object and a preProcess object
}
\description{
almost identical to \code{caret::createModel}
}

