% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getDefaultModel.R
\name{getDefaultModel}
\alias{getDefaultModel}
\title{list of list of models}
\usage{
getDefaultModel(n = Inf, y, type = c(NA, "classification", "regression",
  "survival"))
}
\arguments{
\item{n}{number of models to list}

\item{y}{response vector, used to determine problem type if \code{type} is missing}

\item{type}{type of problem. Default to be determined by \code{y}}
}
\value{
A list of model names
}
\description{
list of list of models
}

