% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getModelInfo.R
\name{getModelInfo}
\alias{getModelInfo}
\title{get model information}
\usage{
getModelInfo(model = NULL, regex = TRUE, ...)
}
\arguments{
\item{model}{Name of model}

\item{regex}{Should regular expression be used for matching}

\item{...}{Other argument passed to \code{\link{grepl}}}
}
\value{
A list of matched models, each of which is a list of fit, predict methods
}
\description{
search built-in model informations, such as fit, predict methods. It is also identical to \code{caret::getModelInfo}
}
\seealso{
\code{caret::getModelInfo}
}

