% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{lazyML}
\alias{lazyML}
\alias{lazyML-package}
\title{Automatic machine learning algorithms selection and hyper-parameters optimization.}
\description{
This package is built upon the caret package my Max Kuhn, with auto algorithm selection and hyper-parameter optimization.
}
\details{
The package is designed to use a single call to \code{\link{mpTune}} to get model selection and hyper-parameter optimization done at once,
with optional parallel choice. Other possible usage is to first use \code{\link{createCVFolds}} to create cross validation folds, and use
multiple calls to \code{\link{mpTune}} with different models using 'qsub', and use \code{\link{combine.mpTune}} to combine all the results.
}

