% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loopingRule.R
\name{loopingRule}
\alias{foreachLoop}
\alias{lapplyLoop}
\alias{loopingRule}
\alias{mclapplyBatchLoop}
\alias{mclapplyLoop}
\alias{parLapplyBatchLoop}
\alias{parLapplyLBLoop}
\alias{parLapplyLoop}
\title{parallel looping functions}
\usage{
parLapplyBatchLoop(executeFunciton, loopList, ..., cl, batchSize = 5,
  mc.cores = 1, shuffle = TRUE)

parLapplyLoop(executeFunciton, loopList, ..., cl, shuffle = TRUE)

parLapplyLBLoop(executeFunciton, loopList, ..., cl)

mclapplyBatchLoop(executeFunciton, loopList, ...,
  parNumber = getOption("mc.cores"), batchSize = 2, preschedule = FALSE)

mclapplyLoop(executeFunciton, loopList, ...,
  parNumber = getOption("mc.cores"), preschedule = FALSE)

lapplyLoop(executeFunciton, loopList, ..., parNumber = getOption("mc.cores"),
  preschedule = FALSE)

foreachLoop(executeFunciton, loopList, ...)
}
\arguments{
\item{executeFunciton}{An internal task execution function}

\item{loopList}{An internal model list to execute over}

\item{...}{Other parameter to pass to \code{executeFunciton}. Must be left in a loopingRue}

\item{cl}{A SNOW cluster as decribed in \code{parallel::parLapplyLoop}}

\item{batchSize}{Number of models to send to a cluster or core at a time}

\item{mc.cores}{Number of cores / CPUs to use in each compute node}

\item{shuffle}{Should loopList be shuffle before depatch to different compute nodes. Shuffle the loopList
gives a more balanced batch load}

\item{parNumber}{Number of cores / CPUs to use for parallel. See \code{parallel::mclapply} for detail}

\item{preschedule}{Should models depatch be done before sent to nodes or cores. See \code{parallel::mclapply} for detail}
}
\description{
parallel looping functions, usually called by \code{\link{modifyFunction}} to pass parameter, for example \code{cl} in parLapplyLoop
}

