% Generated by roxygen2 (4.0.1): do not edit by hand
\name{metric}
\alias{availableMetric}
\alias{defaultSummary}
\alias{metric}
\alias{multiClassSummary}
\alias{regressionSummary}
\alias{requireSummary}
\alias{survivalSummary}
\alias{twoClassSummary}
\title{performance metrics}
\usage{
twoClassSummary(data, lev = levels(as.factor(data[, "obs"])), model = NULL,
  metric = availableMetric("twoClass"), ...)

multiClassSummary(data, lev = levels(as.factor(data[, "obs"])),
  model = NULL, metric = availableMetric("multiClass"), ...)

regressionSummary(data, lev = NULL, model,
  metric = availableMetric("regression")[1:4], ...)

survivalSummary(data, lev, model, is.prob = NA, surv.percentile = 0.5,
  max.surv.time = max(data$obs[, 1], na.rm = TRUE),
  metric = availableMetric("survival")[1:3], na.rm = TRUE, ...)

defaultSummary(data, ..., type = c(NA, "twoClass", "multiClass", "regression",
  "survival"))

requireSummary(...)

availableMetric(type = c("twoClass", "multiClass", "regression", "survival"),
  is.prob = TRUE)
}
\arguments{
\item{data}{Data frame with columns 'obs', 'pred' and possible columns with class labels as names if class probabilities needed}

\item{metric}{Performance metrics to be computed}

\item{type}{Problem type, either 'twoClass', 'multiClass', 'regression' or 'survival'}

\item{is.prob}{TRUE or FALSE indicated if class probability is availabed in \code{data}.}

\item{lev}{Class labels for a classification problem. NULL for others}

\item{model}{Model names (optional)}

\item{surv.percentile}{Percentile used to compute quantile survival time as predition if data has time to survival information.}

\item{max.surv.time}{Maximum follow-up time}

\item{na.rm}{Should NA rows in data be removed before metric calculation}

\item{...}{Other argument to pass to internal methods}
}
\description{
built-in performance metrics for classification, regression and survival
}
\details{
\code{availableMetric} gives a list of metrics currently supported. \code{defaultSummary} is a wrap over \code{twoClassSummary},
\code{multiClass}, \code{regressionSummary} and \code{survivalSummary}. It automatically determines which one to use by response
in \code{data$obs}. \code{requireSummary} a variation of \code{\link{modifyFunction}} to modify \code{defaultSummary},
particularly the \code{metric} argument to request a vector of non-default metric. See \code{\link{modifyFunction}} for example
}

