% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{modifyFunction}
\alias{modifyFunction}
\title{modify default arguments of a function}
\usage{
modifyFunction(f, ...)
}
\arguments{
\item{f}{A function to be modified}

\item{...}{Arguments an values key pairs}
}
\value{
A function with modified default values
}
\description{
modify default arguments of a function
}
\examples{
# function twoClassSummary has an argument 'metric' which defaults to long vector of metrics,
# to make a numer twoClassSummary function that returns only 'Kappa' and 'AUC', we can do
newTwoClassSummary <- modifyFunction(twoClassSummary, metric = c('Kappa', 'AUC'));
}

