% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/more_mpTune.R
\name{more.mpTune}
\alias{more.mpTune}
\title{tune more models}
\usage{
\method{more}{mpTune}(obj, x = obj$data[[1]], y = obj$data[[2]],
  weights = obj$data$weights, models, ...,
  gridLength = obj$config$gridLength[seq_along(models)],
  randomizedLength = obj$config$randomizedLength[seq_along(models)])
}
\arguments{
\item{obj}{A 'mpTune' object}

\item{x}{A design matrix or a formula}

\item{y}{A response vector or a data.frame, which must be consistent with \code{x}}

\item{weights}{Sample weigth}

\item{models}{A vector of names from \code{\link{models}} in models.RData}

\item{...}{Other arguments (expect 'mpTnControl') passed to 'mpTune'}

\item{gridLength}{Grid length if using grid search for hyper-parameters (tuning parameters)}

\item{randomizedLength}{Number of hyper-parameter configuration if randomizedSearch is available for models}
}
\value{
A 'mpTune' object with more models
}
\description{
tune more models using the same resampling setting and performance metric for fair comparison
}
\seealso{
See \code{\link{mpTune}} for examples
}

