% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mpTuneControl.R
\name{mpTuneControl}
\alias{mpTuneControl}
\title{Generate mpTnControl for mpTune}
\usage{
mpTuneControl(samplingFunction = createCVFolds, ..., sampleIndex = NULL,
  problemType = NA, classProbs = TRUE, preProcess = NULL,
  summaryFunction = defaultSummary, returnData = TRUE,
  allowParallel = TRUE)
}
\arguments{
\item{samplingFunction}{A function to takes x, y as input and generate a list of training indices of cases (rows in x)}

\item{...}{Other argument to be passed to \code{samplingFunction}}

\item{sampleIndex}{A list of training indices. If given, will be used directly and \code{samplingFunction} will be ignored}

\item{problemType}{'classification', 'regression' or 'survival'. Default to be automatically determined. Currently not effective any way}

\item{classProbs}{Should class probabilities be returned in prediction}

\item{preProcess}{A caret preProcess function}

\item{summaryFunction}{summary function for model performance. See caret;;defaultSummary for detail of how to customize. Default to defaultSummary which
automatically guess problemType from input and determine what to return}

\item{returnData}{Should data be saved in the mpTune output}

\item{allowParallel}{Should models be tuned in parallel}
}
\value{
A list ready for \code{mpTnControl} in \code{\link{mpTune}}.
}
\description{
Generate mpTnControl for mpTune
}

