% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survival_misc.R
\name{predict.coxph}
\alias{predict.coxph}
\title{predict coxph model}
\usage{
\method{predict}{coxph}(object, newdata, type = c("lp", "risk", "expected",
  "terms", "median", "mean", "prob"), max.surv.time = max(object$y[, 1]), ...)
}
\arguments{
\item{object}{Coxph object}

\item{newdata}{Newdata to predict}

\item{type}{Type of prediction}

\item{max.surv.time}{Usually maximum follow-up time}

\item{...}{Other argument passed to survival :  : predict.coxph}
}
\value{
a vector of predicted value or a matrix of survival probability if type = 'prob', with colnames = times
}
\description{
extend survival::predict.coxph to give 'median', 'mean' and a vector a survial quantile
}

