% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.mpTune.R
\name{print.mpTune}
\alias{print.mpTune}
\title{print mpTune result}
\usage{
print.mpTune(x, metric = x$performanceMetric[1], digits = 4,
  row.names = FALSE, ...)
}
\arguments{
\item{x}{mpTune object}

\item{metric}{Performance metric to determine the best model. Default to the 1st one}

\item{digits}{Digits to keep for numeric result}

\item{row.names}{Should row name / number be printed}

\item{...}{Other arguments pass to the default base print function}
}
\description{
Print information of the best model, failure and message if any
}

