% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/resample_mpTune.R
\name{resample}
\alias{resample}
\alias{resample.mpTune}
\alias{resample.mpTune.formula}
\title{Resample to evaluate performance of \code{\link{mpTune}} chosen model}
\usage{
resample(obj, ...)

\method{resample}{mpTune.formula}(obj, formula = obj$data$formula,
  data = obj$data$data, ...)

\method{resample}{mpTune}(obj, x = obj$data$x, y = obj$data$y,
  weights = obj$data$weights, samplingFunction = createCVFolds, ...,
  returnAllSummaries = TRUE, summaryArgs = NULL, returnAllMpTunes = FALSE,
  printresampleInfo = TRUE, parallelResampling = FALSE)
}
\arguments{
\item{obj}{A \code{\link{mpTune}} object returned from a call to \code{\link{mpTune}}}

\item{...}{Further arguments to be passed to \code{\link{resampling}}}

\item{formula}{A model formula. Default to obj$data$formula if any}

\item{data}{Data to fit. Default to obj$data$data if any}

\item{x}{Design matrix if not in \code{obj}}

\item{y}{Response vector if not in \code{obj}}

\item{weights}{Sample weights if not in \code{obj}}

\item{samplingFunction}{Any function that returns a list of indices for training}

\item{returnAllSummaries}{Should all \code{summarizedMpTune} be returned}

\item{summaryArgs}{Additional argument passed to \code{\link{summary.mpTune}}}

\item{returnAllMpTunes}{Should all \code{mpTune} objects be returned}

\item{printresampleInfo}{If to print resampling fitting progress (not implemented yet)}

\item{parallelResampling}{Should resampling fitting be parallelled (not implemented yet)}
}
\value{
A 'resampledMpTune' object with entries: \cr
		$ aggregatePerformance: a data.frame gives an aggregated (over resamples) performance \cr
		$ allResamplePerformance: a data.frame gives detail performance on earch resample \cr
		$ sampleIndex: a list of resample indices for training \cr
}
\description{
Resample (e.g., cross validation, cross validation) to estimate the true performance
of the whole \code{\link{mpTune}} model choosing precedure.
}
\seealso{
See \code{\link{mpTune}} for examples
}

