% Generated by roxygen2 (4.0.1): do not edit by hand
\name{resampling}
\alias{createBootstrap}
\alias{createCVFolds}
\alias{resampling}
\title{create (repeated) cross validation folds}
\usage{
createCVFolds(x, y, nfold = 5, repeats = 1, stratify = FALSE,
  seed = NULL)

createBootstrap(x, y, n, repeats = 25, replace = TRUE, stratify = FALSE,
  seed = NULL)
}
\arguments{
\item{x}{Design matrix}

\item{y}{Response vector}

\item{nfold}{Number of folds}

\item{repeats}{Number of repeats}

\item{stratify}{Should folds be stratified on response classes or censoring indicators (if survival)}

\item{seed}{Seed to set before sampling}

\item{replace}{If to sample with replacement}

\item{n}{Bootstrapped sample size (if integer) or proportional}
}
\value{
A list of repeated cross validation training samples (as row numbers of x) or a list of Bootstrap samples (as row numbers of x as well).
}
\description{
create (repeated) cross validation folds
}

