% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SCI}
\alias{SCI}
\title{A Smooth concordance loss function for mboost algorithms}
\usage{
SCI(alpha = 1)
}
\arguments{
\item{alpha}{the scale parater in 1/(1+alpha*z) to approach the function with value = {0, 1}. Default to 1}
}
\value{
a \code{mboost::Family} object for mboost algorithms
}
\description{
A Smooth concordance loss function for mboost algorithms
}

