% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{sigest.random}
\alias{addInteraction}
\alias{sigest.random}
\title{Interal functions}
\usage{
sigest.random(x, frac = 0.5, scaled = TRUE, na.action = na.omit,
  len = 3)

addInteraction(x, squared = TRUE)
}
\arguments{
\item{x}{Design matrix}

\item{frac}{Fraction of data used}

\item{scaled}{If to scale column in x}

\item{na.action}{How to handle missing values}

\item{len}{Number of sigma inverse values to general}

\item{squared}{If squared terms should be added}
}
\value{
A vector of of sigma inverse values
}
\description{
Interal functions
}
\details{
\code{sigest.random} is a modified version of \code{kernlab::sigest}, aiming to generate
tuning parameters for sigmal inverse of an RBF kernel, base on \code{kernlab::sigest}. It returns
a vector of sigma inverse value. For details, see \code{kernlab::sigest}.
\code{addInteraction} adds 2-way interactions and/or squared terms to the design matrix.
}
\seealso{
Please see kernlab::sigest for details
}

