% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary_mpTune.R
\name{summary.mpTune}
\alias{print.summarizedMpTune}
\alias{summary.mpTune}
\title{summarize result from \code{\link{mpTune}}}
\usage{
\method{summary}{mpTune}(object, metric = object$performanceMetric,
  maximize = sapply(metric, function(x) ifelse(x \%in\% c("RMSE", "MAE"),
  FALSE, TRUE)), by = c("metric", "model"), ...)

\method{print}{summarizedMpTune}(x, indent = rep(" ", 4), prefix = c("-",
  "-", "+"), digits = 3, justify = c("left", "right", "center", "none"),
  ...)
}
\arguments{
\item{object}{A 'mpTune' object or a 'summarizedMpTune' object}

\item{metric}{A vector a metric for sumary. Default to metrics used in \code{obj}}

\item{maximize}{A logical vector inddicating if metric values the larger the better}

\item{by}{Which should be used for the first level of the returned list. See the following for detail}

\item{...}{Just for compatibility}

\item{x}{A 'mpTune' object or a 'summarizedMpTune' object}

\item{indent}{A string used to indent he next level of a list}

\item{prefix}{A vector of characters put in front of each entry}

\item{digits}{Number of effective digits to keep for double numeric value}

\item{justify}{Either 'left' (default), 'right', 'center', 'none' for alignment of keys (list names)}
}
\value{
summary returns a 'summarizedMpTune' object, which is a list of two levels, either
performance metrics -> models (by = 'metric') or models -> performance metrics, where models are always ranked by
the performance
}
\description{
summarize a \code{\link{mpTune}} result and print summary information
}
\seealso{
See \code{\link{mpTune}} for examples
}

