% Generated by roxygen2 (4.0.1): do not edit by hand
\name{survival.quantiles}
\alias{survival.quantiles}
\title{get survival quantiles}
\usage{
survival.quantiles(surv.percentile = 0.5, prediction, times,
  max.surv.time = max(times))
}
\arguments{
\item{surv.percentile:}{a vector of survival probabilities.}

\item{prediction:}{a matrix of survival probilites, with row = sample, col = time}

\item{times:}{a vector of times of interest}

\item{max.surv.time:}{Usually maximum follow-up time}
}
\value{
a matrix of survival quantitles (times), with row = sample, col = surv.percentile
}
\description{
get survival quantiles from survival curves for each sample
}

