% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_mpTune.R
\name{fit}
\alias{fit}
\alias{fit.mpTune}
\alias{fit.mpTune.formula}
\title{fit the best model according to \code{\link{mpTune}}}
\usage{
fit(obj, ...)

\method{fit}{mpTune.formula}(obj, formula = obj$data$formula,
  data = obj$data$data, ...)

\method{fit}{mpTune}(obj, x = obj$data$x, y = obj$data$y,
  weights = obj$data$weights, model = 1, ...)
}
\arguments{
\item{obj}{A \code{\link{mpTune}} object returned from a call to \code{\link{mpTune}}}

\item{...}{Further arguments to be passed to \code{\link{summary.mpTune}}}

\item{formula}{Model formula if not in \code{obj}}

\item{data}{Ata if not in \code{obj}}

\item{x}{Design matrix if not in \code{obj}}

\item{y}{Response vector if not in \code{obj}}

\item{weights}{Sample weights if not in \code{obj}}

\item{model}{An interger or model name indicating which model to fit. Default to 1, the best model}
}
\value{
A 'bestMpTuneModel' or 'bestMpTuneModel.formula' object with entries: \cr
        $ fit: model fit \cr
        $ preProc \cr
        $ modelName \cr
        $ modelInfo: a list of models fitting, predict information, as from caret \cr
        $ selectionMetric: the performace matric used to select the best model \cr
        $ mpTunePerformance: performance of best model in the mpTune object \cr
}
\description{
fit the best model according to \code{\link{mpTune}}
}
\seealso{
See \code{\link{mpTune}} for examples
}

