% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/more_mpTune.R
\name{more}
\alias{more}
\title{tune more models or do more resampling}
\usage{
more(obj, ...)
}
\arguments{
\item{obj}{A 'mpTune' or 'resampledMpTune'}

\item{...}{Other arguments passed to \code{\link{mpTune}}}
}
\value{
same object as \code{x}
}
\description{
tune more models using the same resampling setting and performance metric for fair comparison
}

