% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survival_misc.R
\name{SCI}
\alias{SCI}
\title{A Smooth concordance loss function for mboost algorithms}
\usage{
SCI(alpha = 1)
}
\arguments{
\item{alpha}{the scale parater in 1/(1+alpha*z) to approach the function with value = {0, 1}. Default to 1}
}
\value{
a \code{mboost::Family} object for mboost algorithms
}
\description{
A Smooth concordance loss function for mboost algorithms
}
\references{
http://www.hindawi.com/journals/cmmm/2013/873595/
}

