% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survival_misc.R
\name{survival.quantiles}
\alias{survival.quantiles}
\title{get survival quantiles}
\usage{
survival.quantiles(surv.percentile = 0.5, prediction, times,
  max.surv.time = max(times))
}
\arguments{
\item{surv.percentile}{A vector of survival probabilities.}

\item{prediction}{A matrix of survival probilites, with row = sample, col = time}

\item{times}{A vector of times of interest}

\item{max.surv.time}{Usually maximum follow-up time}
}
\value{
A matrix of survival quantitles (times), with row = sample, col = surv.percentile
}
\description{
Get survival quantiles from survival curves for each sample
}

